Split up browse_games functions into separate specialized files that can be used by other pages
