+++ client/src/browse_games.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="UTF-8" />
         <title>Gamer Port - Recommend Games to Friend in a Minute</title>
-        <script src="js/browse_games.js" defer></script>
+        <script type="module" src="js/browse_games.js" defer></script>
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
         <link rel="stylesheet" href="https://pro.fontawesome.com/releases/v5.10.0/css/all.css" integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p" crossorigin="anonymous"/>
         <link rel="stylesheet" href="css/browse_games.css">
