+++ client/src/js/rating.js
@@ -1,0 +1,61 @@
+'use strict';
+const url = 'http://localhost:8080';
+const userID = '1111';
+
+// Submits a rating for a game to the server 
+export async function ratingSubmit(ratingsDiv, gameID) {
+    let starCount = 0;
+    for (let i = 1; i <= 5; i++) {
+        if (ratingsDiv.childNodes[i].style.color === 'gold') {
+            starCount++;
+        }
+        if (starCount > 0) {
+            const response = await fetch(url+'/user/ratings/update', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({'userID':userID,'rating':starCount,'gameID':gameID})
+            });
+            if (!response.ok) {
+                throw "Error adding rating to ratings list";
+            } 
+        } else {
+            const response = await fetch(url+'/user/ratings/remove', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({'userID':userID,'gameID':gameID})
+            });
+            if (!response.ok) {
+                throw "Error removing rating from ratings list";
+            } 
+        }
+    }
+}
+
+// Function for selecting/de-selecting stars for rating
+export function clickStar(starDiv, ratingsDiv, starCount) {
+
+    if (starDiv.style.color === 'gold' && (starCount === 5 || ratingsDiv.childNodes[starCount+1].style.color !== 'gold')) {
+        for (let i = starCount; i >= 1; i--) {
+            if (ratingsDiv.childNodes[i].style.color === 'gold') {
+                ratingsDiv.childNodes[i].style.color = 'black';
+            }
+        }
+        return;
+    }
+
+    for (let i = starCount; i >= 1; i--) {
+        if (ratingsDiv.childNodes[i].style.color !== 'gold') {
+            ratingsDiv.childNodes[i].style.color = 'gold';
+        }
+    }
+    
+    for (let i = starCount + 1; i <= 5; i++) {
+        if (ratingsDiv.childNodes[i].style.color === 'gold') {
+            ratingsDiv.childNodes[i].style.color = 'black';
+        }
+    }
+}
